refactor: Abstracted Accordion into it's own component
